smartmeshsdk3/requirements.txt
pyserial
requests
paho-mqtt
